+++ apps/native/app/_layout.tsx
@@ -65,7 +65,6 @@ export default function Layout() {
               <HeroUINativeProvider>
                 <Stack>
                   <Stack.Protected guard={isLoggedIn && hasCompletedOnboarding}>
-                    <Stack.Screen name="modal" options={{ presentation: "modal" }} />
                     <Stack.Screen name="(tabs)" options={{ headerShown: false }} />
                   </Stack.Protected>
