+++ apps/native/deprecated/_layout.tsx
@@ -54,7 +54,6 @@ export default function Layout() {
                     <Stack.Screen name="index" options={{ headerShown: false }} />
                     <Stack.Screen name="chat" options={{ headerShown: false }} />
                     <Stack.Screen name="test/ai-stream" options={{ headerShown: false }} />
-                    <Stack.Screen name="modal" options={{ presentation: "modal" }} />
                     <Stack.Screen name="(tabs)" options={{ headerShown: false }} />
                     <Stack.Screen name="planners/create" options={{ headerShown: false }} />
                   </Stack.Protected>
