+++ packages/db/src/queries/index.ts
@@ -1,7 +1,6 @@
 export * from "./plan-drafts";
 export * from "./monthly-plans";
 export * from "./plan-tasks";
-export * from "./user-goals";
 export * from "./calendar";
 export * from "./progress";
 export * from "./plan-generation";
@@ -14,3 +13,4 @@ export * from "./reminders";
 export * from "./pattern-detection";
 export * from "./yearly-resolutions";
 export * from "./user-preferences";
+export * from "./user";
